requirements.txt
ibis-framework[duckdb,polars,deltalake] @ git+https://github.com/ibis-project/ibis.git@master
matplotlib
seaborn
plotly
altair @ git+https://github.com/altair-viz/altair.git@master
streamlit

